Minor folder refactor + added launcher icon + fixed app name shown
