+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:name=".EntranceActivity"
             android:configChanges="orientation|keyboardHidden|screenSize"
             android:exported="true"
-            android:label="@string/title_activity_entrance"
+            android:label="@string/app_name"
             android:theme="@style/Theme.Vinilos.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
