+++ app/src/main/java/com/example/vinilos/ui/adapters/AlbumAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.vinilos.adapters
+package com.example.vinilos.ui.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
