+++ app/src/main/java/com/example/vinilos/ui/adapters/BindingAdapterUtils.kt
@@ -1,4 +1,4 @@
-package com.example.vinilos.adapters
+package com.example.vinilos.ui.adapters
 
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
