+++ app/src/main/java/com/example/vinilos/viewmodels/CollectorsViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.vinilos.ui.viewmodels
+package com.example.vinilos.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
