+++ app/src/main/java/com/example/vinilos/viewmodels/GuestViewModel.kt
@@ -1,16 +1,16 @@
-package com.example.vinilos.ui.viewmodels
+package com.example.vinilos.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class CollectorLandingViewModel : ViewModel() {
+class GuestViewModel : ViewModel() {
 
     private val _title = MutableLiveData<String>().apply {
-        value = "¡Bienvenido amigo coleccionista!"
+        value = "Bienvenido Invitado"
     }
     private val _content = MutableLiveData<String>().apply {
-        value = "Para acceder a las distinas opciones, haz clic en el menú en la parte superior izquierda"
+        value = "Bienvenido! Para acceder a las distinas opciones del app, haz click en el menu en la parte superior izquierda"
     }
     val title: LiveData<String> = _title
     val content: LiveData<String> = _content
